pyproject.toml
[project]
name = "buchhaltung"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0"
]
